requirements.txt
python-dotenv
python-telegram-bot==13.15
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
